fix(providers): harden runtime and antigravity discovery
Add a tiny Node runtime launcher so unsupported Node versions fail with an actionable upgrade message before terminal dependencies load.

Warn when SQLite-backed provider data exists but node:sqlite cannot load, and tighten Cursor/Goose/OpenCode discovery around that diagnostic.

Teach Antigravity to discover Windows extension-server flags and wrapped RPC payloads, with focused tests for process parsing and runtime guard behavior.
