+++ src/providers/cursor.ts
@@ -4,7 +4,7 @@ import { homedir } from 'os'
 
 import { calculateCost } from '../models.js'
 import { readCachedResults, writeCachedResults } from '../cursor-cache.js'
-import { isSqliteAvailable, getSqliteLoadError, openDatabase, type SqliteDatabase } from '../sqlite.js'
+import { isSqliteAvailable, getSqliteLoadError, openDatabase, warnIfSqliteUnavailable, type SqliteDatabase } from '../sqlite.js'
 import type { Provider, SessionSource, SessionParser, ParsedProviderCall } from './types.js'
 
 const CURSOR_COST_MODEL = 'claude-sonnet-4-5'
@@ -435,10 +435,9 @@ export function createCursorProvider(dbPathOverride?: string): Provider {
     },
 
     async discoverSessions(): Promise<SessionSource[]> {
-      if (!isSqliteAvailable()) return []
-
       const dbPath = dbPathOverride ?? getCursorDbPath()
       if (!existsSync(dbPath)) return []
+      if (warnIfSqliteUnavailable('Cursor')) return []
 
       return [{ path: dbPath, project: 'cursor', provider: 'cursor' }]
     },
