+++ tsup.config.ts
@@ -1,8 +1,10 @@
 import { defineConfig } from 'tsup'
 
 export default defineConfig({
-  entry: ['src/cli.ts'],
+  entry: ['src/cli.ts', 'src/main.ts'],
   format: ['esm'],
+  // Keep this below the runtime floor so dist/cli.js can print the upgrade
+  // message on older Node before dynamically importing the main bundle.
   target: 'node20',
   outDir: 'dist',
   clean: true,
